requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
bitly-api==0.3
bs4==0.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
decorator==4.3.0
fake-useragent==0.1.11
future==0.17.1
idna==2.8
lxml==4.2.5
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.21.0
selenium==3.141.0
six==1.12.0

urllib3==1.24.1
